Add `oclif_pack_from_verdaccio_storage`.
This does roughly what the name suggests: This tries doing an `oclif` pack
from the Verdaccio server that's running locally.  Any packages which are
available locally will be used in the packing, and any which are not there
will be fetched from npm.

When used with the existing Verdaccio publish steps, this should allow the
packed tarball to be precisely what was contained within this commit.