Skip local packages when proxying.
By explicitly skipping the packages which are present in the repository that
the step is being invoked on, rather than using a "use local if present"
approach which I'd employed prior to this, we can make sure we fail when the
packages are suspiciously not present when they're intended to be.

Making the situation which I fixed in the [Previous commit], impossible.

[Previous commit]: e7d4d84c161bb73563ba80a5b83b93e7cce1a12f