Create the parent directory, not the actual directory, before copying.
This subtle difference unfortunately resulted in the directory I had
intended on restoring - a step originally introduced in #1 - being restored
to the directory inside this directory, rather than in the exact place of
the intended destination.

A follow up commit will make this more noticeable in terms of recognizing
the result (failure).